+++ cpp/CHANGELOG.md
@@ -20,6 +20,7 @@ Please mark all change in change log and use the ticket from JIRA.
 - MS-359 - Add cost test in new scheduler
 - MS-361 - Add event in resource
 - MS-364 - Modify tasktableitem in tasktable
+- MS-365 - Use tasktableitemptr instead in event
 
 ## New Feature
 - MS-343 - Implement ResourceMgr